feat: 체력 상한 변경에 따른 자료형 수정

+++ ScoreBoard/ScoreBoard/data/stat/Stat.cs
@@ -10,8 +10,8 @@ namespace ScoreBoard.data.stat
 {
     public class Stat
     {
-        public ushort Hp { get; set; } // 체력 (예: 10, 20 등)
-        public ushort MaxHp { get; set; } // 최대 체력 (예: 10, 20 등)
+        public int Hp { get; set; } // 체력 (예: 10, 20 등)
+        public int MaxHp { get; set; } // 최대 체력 (예: 10, 20 등)
         public ushort Shield { get; set; } = 0; // 방어막 (예: 5, 10 등)
         public ushort Movement { get; set; } // 이동 가능 거리 (예: 1, 2, 3 등)
         public ushort? SpellPower { get; set; } // 주문력 (예: 10, 20 등) - 선택적 속성
